♻️ Refactor: 전역 상태 리팩토링
