+++ src/utils/index.js
@@ -1,3 +1,3 @@
 export { generateRandomNumber } from "./generateRandomNumber"
 export { readLineAsync } from "./readLine"
-export { NumberTypeError } from "./error"
+export { useState } from "./createState"
